commit withc checkpoint on 02-03-15

+++ src/main/java/com/projectx/data/controller/completeregister/TransactionUpdatesController.java
@@ -2,10 +2,15 @@ package com.projectx.data.controller.completeregister;
 
 
 
+import java.util.Date;
+
+import javax.validation.Valid;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
@@ -21,7 +26,8 @@ import com.projectx.data.repository.completeregister.TransactionalUpdatesReposit
 import com.projectx.data.repository.completeregister.VendorDetailsCustomRepository;
 import com.projectx.rest.domain.completeregister.CustomerOrVendorDetailsDTO;
 import com.projectx.rest.domain.quickregister.CustomerIdTypeEmailTypeDTO;
-import com.projectx.rest.domain.quickregister.CustomerIdTypeMobileTypeDTO;
+import com.projectx.rest.domain.quickregister.CustomerIdTypeEmailTypeUpdatedByDTO;
+import com.projectx.rest.domain.quickregister.CustomerIdTypeMobileTypeUpdatedByDTO;
 import com.projectx.rest.domain.quickregister.CustomerQuickRegisterEmailMobileVerificationEntity;
 
 @RestController
@@ -38,8 +44,10 @@ public class TransactionUpdatesController {
 	VendorDetailsCustomRepository vendorDetailsCustomRepository;
 
 	@RequestMapping(value="/updateCustomerDetails",method=RequestMethod.POST)
-	public ResponseEntity<CustomerDetails> updateCustomerDetails(@RequestBody CustomerDetails customerDetails)
+	public ResponseEntity<CustomerDetails> updateCustomerDetails(@Valid @RequestBody CustomerDetails customerDetails,BindingResult bindingResult)
 	{
+		if(bindingResult.hasErrors())
+			return new ResponseEntity<>(HttpStatus.NOT_ACCEPTABLE);
 		
 		ResponseEntity<CustomerDetails> result=null;
 		CustomerDetails updatedEntity=null;
@@ -50,66 +58,69 @@ public class TransactionUpdatesController {
 		}
 		catch(DataIntegrityViolationException e)
 		{
-			//return customerDetailsCustomRepository.findOne(customerDetails.getCustomerId());
-			result=new ResponseEntity<CustomerDetails>(HttpStatus.NOT_MODIFIED);
+			result=new ResponseEntity<CustomerDetails>(HttpStatus.ALREADY_REPORTED);
 		}
 		return result;
 		
 	}
 	
 	@RequestMapping(value="/updateVendorDetails",method=RequestMethod.POST)
-	public ResponseEntity<VendorDetails> updateVendorDetails(@RequestBody VendorDetails vendorDetails)
+	public ResponseEntity<VendorDetails> updateVendorDetails(@Valid @RequestBody VendorDetails vendorDetails,BindingResult bindingResult)
 	{
-		VendorDetails updatedEntity=null;
+		if(bindingResult.hasErrors())
+			return new ResponseEntity<>(HttpStatus.NOT_ACCEPTABLE);
+		
 		ResponseEntity<VendorDetails> result=null;
-		
-		
 		try{		
-			
-			updatedEntity=transactionalUpdatesRepository.updateVendorDetails(vendorDetails);
+		
+			VendorDetails updatedEntity=transactionalUpdatesRepository.updateVendorDetails(vendorDetails);
 			
 			result=new ResponseEntity<VendorDetails>(updatedEntity, HttpStatus.OK);
 			
-			
 		}catch(DataIntegrityViolationException e)
 		{
-			result=new ResponseEntity<>(HttpStatus.NOT_MODIFIED);
+			result=new ResponseEntity<>(HttpStatus.ALREADY_REPORTED);
 		}
 		
 		return result;
 	}
 	
 	@RequestMapping(value="/updateMobileInDetailsEnityAndAuthenticationDetails",method=RequestMethod.POST)
-	public ResponseEntity<Boolean> updateMobileInDetailsEnityAndAuthenticationDetails(@RequestBody CustomerIdTypeMobileTypeDTO customerIdTypeMobileTypeDTO)
+	public ResponseEntity<Boolean> updateMobileInDetailsEnityAndAuthenticationDetails(@Valid @RequestBody CustomerIdTypeMobileTypeUpdatedByDTO customerIdTypeMobileTypeDTO,
+			BindingResult bindingResult)
 	{
-		Boolean updatedStatus=null;
+		if(bindingResult.hasErrors())
+			return new ResponseEntity<>(HttpStatus.NOT_ACCEPTABLE);
 		
 		ResponseEntity<Boolean> result=null;
 		
 		try{
-			updatedStatus=transactionalUpdatesRepository
+			Boolean updatedStatus=transactionalUpdatesRepository
 				.updateMobileInDetailsEnityAndAuthenticationDetails(customerIdTypeMobileTypeDTO.getCustomerId(), customerIdTypeMobileTypeDTO.getCustomerType(),
-						customerIdTypeMobileTypeDTO.getMobileType());
+						customerIdTypeMobileTypeDTO.getMobileType(),customerIdTypeMobileTypeDTO.getUpdatedBy());
 			
 			result=new ResponseEntity<Boolean>(updatedStatus, HttpStatus.OK);
 			
 		}catch(DataIntegrityViolationException e)
 		{
-			result=new ResponseEntity<>(HttpStatus.NOT_MODIFIED);
+			result=new ResponseEntity<>(HttpStatus.ALREADY_REPORTED);
 		}
 		return result;
 	}
 	
 	@RequestMapping(value="/updateEmailInDetailsEnityAndAuthenticationDetails",method=RequestMethod.POST)
-	public ResponseEntity<Boolean> updateEmailInDetailsEnityAndAuthenticationDetails(@RequestBody CustomerIdTypeEmailTypeDTO customerIdTypeEmailTypeDTO)
+	public ResponseEntity<Boolean> updateEmailInDetailsEnityAndAuthenticationDetails(@Valid @RequestBody CustomerIdTypeEmailTypeUpdatedByDTO customerIdTypeEmailTypeDTO,
+			BindingResult bindingResult)
 	{
-		Boolean updatedStatus=null;
+		if(bindingResult.hasErrors())
+			return new ResponseEntity<>(HttpStatus.NOT_ACCEPTABLE);
+		
 		ResponseEntity<Boolean> result=null;
 		
 		try{
-		 updatedStatus=transactionalUpdatesRepository
+			Boolean updatedStatus=transactionalUpdatesRepository
 				.updateEmailInDetailsEnityAndAuthenticationDetails(customerIdTypeEmailTypeDTO.getCustomerId(), customerIdTypeEmailTypeDTO.getCustomerType(),
-						customerIdTypeEmailTypeDTO.getEmailType());
+						customerIdTypeEmailTypeDTO.getEmailType(),customerIdTypeEmailTypeDTO.getUpdatedBy());
 		 result=new ResponseEntity<Boolean>(updatedStatus, HttpStatus.OK);
 		 
 		}catch(DataIntegrityViolationException e)
@@ -120,16 +131,17 @@ public class TransactionUpdatesController {
 	}
 	
 	@RequestMapping(value="/saveNewQuickRegisterEntity",method=RequestMethod.POST)
-	public ResponseEntity<CustomerQuickRegisterEmailMobileVerificationEntity> saveNewQuickRegisterEntity(@RequestBody QuickRegisterEntity quickRegisterEntity)
+	public ResponseEntity<CustomerQuickRegisterEmailMobileVerificationEntity> saveNewQuickRegisterEntity(@Valid @RequestBody QuickRegisterEntity quickRegisterEntity,
+			BindingResult bindingResult)
 	{
+		if(bindingResult.hasErrors())
+			return new ResponseEntity<>(HttpStatus.NOT_ACCEPTABLE);
 		
 		ResponseEntity<CustomerQuickRegisterEmailMobileVerificationEntity> result=null;
 		
-		CustomerQuickRegisterEmailMobileVerificationEntity returnEntity=null;
-		
 		try
 		{
-			returnEntity=transactionalUpdatesRepository.saveNewQuickRegisterEntity(quickRegisterEntity);
+			CustomerQuickRegisterEmailMobileVerificationEntity returnEntity=transactionalUpdatesRepository.saveNewQuickRegisterEntity(quickRegisterEntity);
 			result=new ResponseEntity<CustomerQuickRegisterEmailMobileVerificationEntity>(returnEntity, HttpStatus.CREATED);
 			
 		}catch(DataIntegrityViolationException e)
@@ -141,14 +153,17 @@ public class TransactionUpdatesController {
 	}
 	
 	@RequestMapping(value="/deleteQuickRegisterEntityCreateDetails",method=RequestMethod.POST)
-	public ResponseEntity<CustomerOrVendorDetailsDTO> deleteQuickRegisterEntityCreateDetails(@RequestBody QuickRegisterEntity quickRegisterEntity)
+	public ResponseEntity<CustomerOrVendorDetailsDTO> deleteQuickRegisterEntityCreateDetails(@Valid @RequestBody QuickRegisterEntity quickRegisterEntity,
+			BindingResult bindingResult)
 	{
-		ResponseEntity<CustomerOrVendorDetailsDTO> result=null;
-		CustomerOrVendorDetailsDTO customerOrVendorDetailsDTO=null;
+		if(bindingResult.hasErrors())
+			return new ResponseEntity<>(HttpStatus.NOT_ACCEPTABLE);
 		
+		ResponseEntity<CustomerOrVendorDetailsDTO> result=null;
+				
 		try{
 			
-			customerOrVendorDetailsDTO=transactionalUpdatesRepository.deleteQuickRegisterEntityCreateDetailsEntity(quickRegisterEntity);
+			CustomerOrVendorDetailsDTO customerOrVendorDetailsDTO=transactionalUpdatesRepository.deleteQuickRegisterEntityCreateDetailsEntity(quickRegisterEntity);
 			result=new ResponseEntity<CustomerOrVendorDetailsDTO>(customerOrVendorDetailsDTO, HttpStatus.CREATED);
 			
 		}
